requirements.txt
numpy
scipy
gensim
torch
matplotlib
scikit-learn
pandas
hdbscan
sentence-transformers
transformers
evaluate
datasets
Pillow
jiwer
ultralytics
optuna
roboflow
pyyaml 